Use objects more than once
The previous commit limited the number of times an object could be used
to just once. This was done to correct an issue where using an object
more than once would result in an error. In correcting said issue, I
unknowingly broke subsequent calls to Output::loadExtensions().

This commit allows for extension objects to be used as many times
as necessary.