Do not replace editable tag references elsewhere
Use encoding to prevent { and } from being replaced within the contexts
of tag references in the titles and other editable areas of posts,
pages, and tags, as well as their respective primary pages.

This ensures that tag references in editable areas remain as raw
pointers, rather than being otherwise parsed.