Assemble path components in common file
This breaks the path component up into an array, then assembles a new
array that includes a key and a value. Once assembled, the new array is
added to $_GET. If the path were to be foo/bar, it would be converted
to $_GET["foo"] = "bar", for example.

By automatically adding the components to $_GET, the index will be able
to check for the existence of a $_GET value, rather than performing an
array search, then assigning a value to $_GET based off of the search.