Redirect invalid pagination ranges to 404 page
If an incorrect range is submitted, the user will be redirected to the
specified 404 page by default. If the 404 page can not be reached, the
user will instead be redirected to the index.

Range checking is not performed in the getPostBlocks() method, as it is
already handled in the getColumn() method instead.