Consolidate getting epochs into one method
Convert epoch_created into machine-readable format for use in the HTML
datetime attribute. Relative epoch functionality should not be a core
functionality; it should instead be implemented as an extension. Also,
the ability to get the epoch for when a piece of content was edited
should be added at a later time, if at all.